+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### v0.1.39
+- Util : Replace `get_namespace` with `get_mod_info`
+- Prevent using the same namespace as another mod
+
 ### v0.1.38
 - Script : Add `.script_name` as an alias for `.name`
 - Stage : Add safety checking to \*Card-related methods + error for invalid string identifiers